Refactor: 포스트와 포스트 댓글 합침
